pyproject.toml
[tool.poetry]
name = "hinagata"
version = "0.1.0"
description = ""
authors = ["horoiwa <horoiwa195@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.4.1"
scikit-learn = "0.24.*"
lightgbm = "^3.3.2"
ray = {extras = ["tune"], version = "^1.11.0"}
dtreeviz = "^1.3.5"
Boruta = "^0.3"
shap = "^0.40.0"
deap = "^1.3.1"
click = "^8.1.2"
pytz = "^2022.1"
hydra-core = "^1.1.1"
sweetviz = "^2.1.3"
japanize-matplotlib = "^1.1.3"
seaborn = "^0.11.2"
pandas-bokeh = "^0.5.5"

[tool.poetry.dev-dependencies]
jupyterlab = "^3.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
